fix(android): extract connectivity monitoring to injectable NetworkMonitor
The previous approach created ConnectivityManager.NetworkCallback directly
inside ChatViewModel. In JVM unit tests, Android stub constructors throw
RuntimeException("Stub!"), so every test was failing at ViewModel creation.

Fix: move all ConnectivityManager / NetworkCallback logic into a
@Singleton NetworkMonitor class (utils/NetworkMonitor.kt). ChatViewModel
now receives NetworkMonitor via constructor injection and collects its
isOffline StateFlow. In tests, NetworkMonitor is replaced by a simple
mockk that exposes MutableStateFlow(false), so no Android stubs are
touched.
