+++ android/app/src/main/kotlin/org/voxquieta/app/presentation/viewmodels/ChatViewModel.kt
@@ -2,9 +2,6 @@ package org.voxquieta.app.presentation.viewmodels
 
 import android.content.Context
 import android.content.Intent
-import android.net.ConnectivityManager
-import android.net.Network
-import android.net.NetworkCapabilities
 import androidx.core.content.FileProvider
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -28,6 +25,7 @@ import org.voxquieta.app.domain.repositories.ContactRepository
 import org.voxquieta.app.presentation.components.ContactFormState
 import org.voxquieta.app.security.TurnstileManager
 import org.voxquieta.app.utils.LogCollector
+import org.voxquieta.app.utils.NetworkMonitor
 import dagger.hilt.android.lifecycle.HiltViewModel
 import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.coroutines.Dispatchers
@@ -133,6 +131,7 @@ class ChatViewModel @Inject constructor(
     private val translationPreferences: TranslationPreferences,
     private val sessionPreferences: SessionPreferences,
     private val bibleApiService: BibleApiService,
+    private val networkMonitor: NetworkMonitor,
 ) : ViewModel() {
 
     private val _uiState = MutableStateFlow(ChatUiState())
@@ -187,26 +186,6 @@ class ChatViewModel @Inject constructor(
             initialValue = TranslationPreferences.DEFAULT_TRANSLATION,
         )
 
-    private val connectivityManager = context.getSystemService(ConnectivityManager::class.java)
-
-    private fun isNetworkAvailable(): Boolean {
-        val network = connectivityManager.activeNetwork ?: return false
-        val caps = connectivityManager.getNetworkCapabilities(network) ?: return false
-        return caps.hasCapability(NetworkCapabilities.NET_CAPABILITY_INTERNET)
-    }
-
-    private val networkCallback = object : ConnectivityManager.NetworkCallback() {
-        override fun onAvailable(network: Network) {
-            _uiState.update { it.copy(isOffline = false) }
-        }
-        override fun onLost(network: Network) {
-            // Only mark offline when there is truly no active network left.
-            if (!isNetworkAvailable()) {
-                _uiState.update { it.copy(isOffline = true) }
-            }
-        }
-    }
-
     private val _chapterSheetState = MutableStateFlow<ChapterSheetState>(ChapterSheetState.Idle)
     val chapterSheetState: StateFlow<ChapterSheetState> = _chapterSheetState.asStateFlow()
     private var loadChapterJob: Job? = null
@@ -239,14 +218,12 @@ class ChatViewModel @Inject constructor(
         viewModelScope.launch { fetchTranslationsWithRetry() }
         // Fetch book name mappings from the backend with retry.
         viewModelScope.launch { fetchBookNamesWithRetry() }
-        // Set initial connectivity state and monitor changes.
-        _uiState.update { it.copy(isOffline = !isNetworkAvailable()) }
-        connectivityManager.registerDefaultNetworkCallback(networkCallback)
-    }
-
-    override fun onCleared() {
-        super.onCleared()
-        connectivityManager.unregisterNetworkCallback(networkCallback)
+        // Mirror the connectivity state into uiState so the UI can react.
+        viewModelScope.launch {
+            networkMonitor.isOffline.collect { offline ->
+                _uiState.update { it.copy(isOffline = offline) }
+            }
+        }
     }
 
     /**
